// src/app/api/generated/model/evaluacion-previa.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Lo minimo para poder comparar
 */
export interface EvaluacionPrevia { 
    dolorEva?: number;
    evolucion?: string;
    iniciadaEn?: string;
}